Bullets are removed if they travel offscreen. Reduces Lag and memory leaks.
In the game loop, the distance between the bullets and the player is calculated. If this distance is more than 2000 px in the x direction, or more than 1000px in the y direction, then the bullet will be deleted. This may be impletemented at a slower rate, but it is to be tested for any lag in its current implementation